moved match peaks by clustering and linear ssignment into separate file
